feat #65 Hub 모듈 Cors 설정, 엔티티 수정, RabbitMQ 설정 수정

+++ tracky-consumer/src/main/java/kernel360/trackyconsumer/application/dto/CarOnOffRequest.java
@@ -34,7 +34,7 @@ public class CarOnOffRequest {
 	private long lon;             // GPS 경도
 	private String ang;           // 방향
 	private String spd;           // 속도(km/h)
-	private double sum;              // 누적 주행 거리(m)
+	private int sum;              // 누적 주행 거리(m)
 
 	public LocationEntity toLocationEntity() {
